move functions out of action file and into utils/filesystem.go
