+++ utils/filesystem.go
@@ -12,14 +12,21 @@
 package utils
 
 import (
+	"archive/zip"
+	"context"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
+	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/eclipse/codewind-installer/errors"
+	"github.com/google/go-github/github"
+	"github.com/urfave/cli"
 	"gopkg.in/yaml.v3"
 )
 
@@ -103,4 +110,93 @@ func PingHealth(healthEndpoint string) bool {
 		log.Fatal("Codewind containers are taking a while to start. Please check the container logs and/or restart Codewind")
 	}
 	return started
+}
+
+//GetZipURL from github api /repos/:owner/:repo/:archive_format/:ref
+func GetZipURL(c *cli.Context) string {
+	branch := c.String("branch")
+	owner := c.String("owner")
+	repo := c.String("repo")
+
+	client := github.NewClient(nil)
+
+	opt := &github.RepositoryContentGetOptions{Ref: branch}
+
+	URL, _, err := client.Repositories.GetArchiveLink(context.Background(), owner, repo, "zipball", opt)
+	if err != nil {
+		log.Fatal(err)
+	}
+	url := URL.String()
+	fmt.Println(URL)
+	return url
+}
+
+//DownloadFile - the git zip of master into current dir
+func DownloadFile(zipFileName, url string) error {
+
+	// Get the data
+	resp, err := http.Get(url)
+	errors.CheckErr(err, 400, "")
+	defer resp.Body.Close()
+
+	// Create the file
+	out, err := os.Create(zipFileName)
+	errors.CheckErr(err, 401, "")
+	defer out.Close()
+
+	// Write body to file
+	_, err = io.Copy(out, resp.Body)
+	fmt.Println(zipFileName)
+
+	return err
+}
+
+//UnZip downloaded file
+func UnZip(zipFileName, destination string) {
+	zipReader, _ := zip.OpenReader(zipFileName)
+
+	var extractedFilePath = ""
+	for _, file := range zipReader.Reader.File {
+
+		zippedFile, err := file.Open()
+		errors.CheckErr(err, 402, "")
+		defer zippedFile.Close()
+
+		fileNameArr := strings.Split(file.Name, "/")
+		extractedFilePath = destination
+
+		for i := 1; i < len(fileNameArr); i++ {
+			extractedFilePath = filepath.Join(extractedFilePath, fileNameArr[i])
+		}
+
+		if file.FileInfo().IsDir() {
+			log.Println("Directory Created:", extractedFilePath)
+			os.MkdirAll(extractedFilePath, file.Mode())
+		} else {
+			log.Println("File extracted:", file.Name)
+
+			outputFile, err := os.OpenFile(
+				extractedFilePath,
+				os.O_WRONLY|os.O_CREATE|os.O_TRUNC,
+				file.Mode(),
+			)
+			errors.CheckErr(err, 403, "")
+			defer outputFile.Close()
+
+			_, err = io.Copy(outputFile, zippedFile)
+			errors.CheckErr(err, 404, "")
+		}
+	}
+	log.Println("File extracted:", zipFileName)
+}
+
+//MoveFiles to directory specified in command
+func MoveFiles(source, destination string) {
+	fmt.Println("==> moving files from ", source)
+	fmt.Println("==> moving files too ", destination)
+
+	err := os.Rename(source, destination)
+	if err != nil {
+		log.Fatal(err)
+	}
 }
